requirements.txt
Flask
selenium
pymysql
mysql-connector